requirements.txt
aiohttp
beautifulsoup4
chromadb
fastapi
feedparser
holidays
httpx
loguru
openai
pandas
psycopg2-binary
pydantic
python-dotenv
pytrends
requests
streamlit
tenacity
uvicorn
